Move cdp build step into an integration test (#80)
* Fix formatting with cargo fmt

* Prevent double layering of generated module name

In the normal Rust module system, a module is represented in the
filesystem by a file of the same name. With the previous setup,
the generated code was saved in `<target_mod>.rs`, with that
file containing another `pub mod <target_mod> { .. }` inside.
Here we remove the outer containing module from the generated code
and instead add it in the cdp crate root, where the generated code
is included, to keep the same paths.

* Commit generated cdp code to the repository

Replace the build script with an integration test that fails if the
generated code is out of date. This has several advantages:

* No need for downstream users to compile the build script
* No need for downstream users to compile the pdl crate
* No need for downstream users to run the code generation step
* The generated code is more usable in an IDE
* CI will fail if the generated code is out of date

The main downside is probably that the generated code is kept in
the repository, but this seems of relatively minimal impact,
especially if the code generator/source files don't change often.