Include Chromium `stderr` when failing to launch (#128)
* breaking: Use async child process

This commit replaces the browser process from `std::process::Child` to `chromiumoxide::async_process::Child`. This new type acts as a wrapper providing compatibility for `async-std` and `tokio` depending on the selected runtime.

Switching to using an async child process allows to use non-blocking I/O on the browser's stderr. This in turns makes it easier to implement error handling on browser start-up.

This is a breaking change since it changes the return type of the public method `Browser::get_mut_child`. This commit focuses only on switching to using an async child process: changes are kept to a minimum and no new behavior is introduced.

* feat: return stderr data on launch error

This commit improves error handling in `Browser::launch`.

1. If any error occurs before the websocket URL could be resolved, the browser `stderr` output is included in the error message.
2. If the browser process exits before the websocket is resolved, an error is returned immediately (without waiting for the launch timeout to expire)

This also fixes a minor error in the previous implementation where checking for `listening on` searched in the full browser output instead of only the last line (small quadratic behavior).

Closes mattsse/chromiumoxide#124